requirements.txt
telebot
google-cloud-bigquery
